+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export * from "./layout";
+export * from "./layouts";
 export * from "./typography";
 export * from "./ui";
